+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `--diff` and `--force` flags can be set during interactive run
+
 ### Fixed
-- Workflows are loaded asynchronously
+- Workflows are downloaded in parallel
 - Avoid loading resources two times in interactive mode
 - Check mode shows a correct command to apply update
 - Workflows were created with non-managed steps sometimes